refactor: Extract SimpleNoopCommand
