+++ lib/plugins/collaborators.js
@@ -1,16 +1,7 @@
 /* eslint-disable camelcase */
 const Diffable = require('./diffable')
 const NopCommand = require('../nopcommand')
-
-class SimpleNoopCommand {
-  constructor (message) {
-    this.message = message
-  }
-
-  toString () {
-    return `SimpleNoopCommand: ${this.message}`
-  }
-}
+const SimpleNoopCommand = require('../simpleNoopCommand')
 
 module.exports = class Collaborators extends Diffable {
   constructor (...args) {
