+++ lib/simpleNoopCommand.js
@@ -1,0 +1,11 @@
+class SimpleNoopCommand {
+  constructor (message) {
+    this.message = message
+  }
+
+  toString () {
+    return `SimpleNoopCommand: ${this.message}`
+  }
+}
+
+module.exports = SimpleNoopCommand
